requirements.txt
appdirs==1.4.0
astroid==1.4.9
backports.functools-lru-cache==1.3
configparser==3.5.0
isort==4.2.5
kafka-python==1.3.2
lazy-object-proxy==1.2.2
mccabe==0.6.1
packaging==16.8
pylint==1.6.5
pyparsing==2.1.10
six==1.10.0
wrapt==1.10.8
